[cfe] Fix bottomType as Never in closures of type schemas
In pre-NNBD Dart the greatest and lowest closures of a type schema were
using `Null` as the bottom type and `dynamic` as the top type. In
NNBD code, the top type should be `Object?', and the bottom type should
be `Never`. Since pre-NNBD code is being deprecated, the code
that allows to choose between the two pairs of extreme types should be
removed, and `Object?` and `Never` should be used as the only options.

For historical reasons, replacing `Object?` with `dynamic` in the
algorithms for the greatest and the lowest closures leads to breakages
in the existing NNBD code. However, the bottom type is assumed to be
`Never` everywhere. This CL removes the ability to parametrise over
the bottom type, choosing `Never` as the only option. The discrepancy
between `Object?` and `dynamic` will be addressed later.

Change-Id: Ia58ccaca37761c5de8e005f11e79b23d02a0e416
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/395963
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>