[analyzer] Rework `InterfaceTypeImpl` constructor to call `NullTypeImpl`
Previously, the `InterfaceTypeImpl._` constructor threw an exception
if an interface type was being constructed for the type `Null`
(because the `NullTypeImpl` constructor should be used instead). This
was silly; as long as we're going to check, we might as well put the
check in the `InterfaceTypeImpl` factory constructor, and have it
simply construct the appropriate subtype. This is simpler, and should
be much lower risk, since we no longer have a code path that has to
throw an exception.

Change-Id: Ie3ed216b285e371866a6a35ccccdf997b8a1f95e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/396620
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>