Language test exercising null shorting inside conditional expressions.
I'm currently working on some refactoring of flow analysis, and I
discovered through a trybot failure that I accidentally broke the
handling of null shorting inside conditional expressions. The trybot
failure was subtle and confusing, though; in order to diagnose it I
had to do a lot of work to reduce it to a straightforward test
case. This CL adds the straightforward test case as a language test,
so that in case the problem recurs, it will be easier to diagnose.

Change-Id: I8a035221d0c9fff3bd0e0cbb8e3b5809b6793726
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/398000
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>