[analysis_server] Only mark positional arguments as editable if they wouldn't require inserting additional values
We can _edit_ all existing arguments because we just replace their values.
We can also _add_ a new argument for a positional parameter as long as there are no missing positional arguments that would need to come before it (because we can't insert them).

Change-Id: Ia66e49f940c463511f472405e90ee46d3b5f6177
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/396841
Reviewed-by: Elliott Brooks <elliottbrooks@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>