[analyzer] Move switch expression resolution into ResolverVisitor.
The analyzer uses two different mechanisms to visit expressions during
resolution:

- `Expression.accept` (which defers to `ResolverVisitor.visit...`
  methods using the standard visitor infrastructure), and

- The abstract method `ExpressionImpl.resolveExpression` (which is
  implemented in every expression type).

The `Expression.accept` approach doesn't support contexts, so it is
only used when the context is the unknown type schema (`_`).

(The reason there are two approaches is largely a historical accident,
and I hope one day to unify them into a single approach. But that's
not the subject of this change.)

Prior to this change, the resolution logic for most expression types
was in the `visit...` method, and the corresponding
`resolveExpression` method did nothing but call `visit...`. But there
was one exception: for switch expressions, the resolution logic was in
`resolveExpression`, and the corresponding `visit...` method called
`analyzeExpression` (which then called `resolveExpression`).

This change moves the resolution logic for switch expressions into the
`visit...` method, and changes the corresponding `resolveExpression`
method to call `visit...`. This makes switch expression resolution
consistent with all other expression types, which should make the
resolution logic easier to reason about.

There should be no functional change.

Change-Id: I209cdb5cd721608f60e6668a4be76fd20a7452de
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/398585
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>