Refactor AnalysisDriver constructor so that more fields can be final.
I think this is a nice win; we just need to make a few utility functions static (and then rename them, and then sort them). And then these fields can be initialized in constructor initializers, and be marked final.

A win for readability and maintainability. We can see at their declarations that the fields are not meant to be mutable.

It is a nice win that the methods can be static as well; they don't
depend on any state of the class, and are just pure calculations made
on their parameters.

Change-Id: Iaeca7dfa80a17a8f81cbb2d07e442fe31c038f46
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/398762
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>