[tests] Enum shorthands - FutureOr tests and some additional tests.
Added back the `FutureOr` tests now that we've decided what we want to do for `FutureOr`. "`FutureOr<S>` denotes the same namespace as `S` for enum shorthands."

There are a few tests for `X extends T` where the context type is `X`. It would be a compile-time error to find static members of `T` even if `T` did have the corresponding static member.

Bug: https://github.com/dart-lang/sdk/issues/57038
Change-Id: Ibb8054d66e024d93b71878ed49bc7866b785b02c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/397986
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Erik Ernst <eernst@google.com>