[linter] Change _isJsInteropType check to check for any instead of just userJsInteropType
EraseNonJSInteropTypes erases extension types except for those
that belong to dart:js_interop. It has an option 'keepUserInteropTypes'
to determine whether to not erase user JS interop types *as well*.
The code before this CL only checked for user interop types and
not dart:js_interop types as well when that flag is enabled.

Note that this doesn't lead to a bug anywhere and therefore there is
no regression test. The only time we use this flag is to analyze the
case where:

1. The right and left types are interop types where the erased left
type (not keeping user interop types) is a subtype of the erased
right type (not keeping user interop types).
2. The right is a user interop type.
3. The erased left type (keeping user interop types) is not a
subtype of the erased right type (keeping user interop types).

If 3 is true given the other conditions, then we lint.

Since the right is a user interop type, the erased right type is
just the right type. The only time this bug could make a difference
is if the left type was a dart:js_interop type that we mistakenly
erased. Since dart:js_interop types and their erased types can
never be a subtype of a user interop type anyways, this bug could
never have given a different value for 3 given the other conditions.

This is therefore purely cleanup.

Change-Id: I6aeaefd652e5f00094b5a89fda2d99b89a9bc516
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/399202
Commit-Queue: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Auto-Submit: Srujan Gaddam <srujzs@google.com>