[analyzer] Narrow the scope over which flow analysis is active.
Change the `ResolverVisitor` so that it creates a `FlowAnalysis`
object just before visiting the initializer or method body of a top
level declaration, and cleans it up just after visiting it.

Previously, the `ResolverVisitor` behaved this way for some top level
declarations, but for others, it created the `FlowAnalysis` object
just before visiting the whole top level declaration, which meant that
flow analysis would be active during resolution of other constructs
that shouldn't participate in flow analysis, such as documentation
comments, metadata, return types, parameters, and type parameters.

There is no functional change now, but in a follow-up CL, I plan to
use the presence or absence of the `FlowAnalysis` object as a debug
check to help make sure the `ResolverVisitor` is visiting
subexpressions in the right way. This change will reduce the need for
weird exceptions in tthat debug check.

Note that the methods `topLevelDeclaration_enter` and
`topLevelDeclaration_exit` (which are responsible for creating and
cleaning up the `FlowAnalysis` object) have been renamed to
`bodyOrInitializer_enter` and `bodyOrInitializer_exit`, to clarify
that they are intended to bracket just the method bodies and
initializers, not whole top level declarations.

Change-Id: Ia2bae84b123bdf3f6893b267393d7dde2a7abeea
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/399631
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>