[analyzer] Fix unnecessary recursive call to ResolverVisitor.resolveForWrite.
When `ResolverVisitor.resolveForWrite` rewrites the LHS of an
assignment from a prefixed identifier to a property access (which it
only does for record types), it's not necessary to make a recursive
call to `resolveForWrite`, because the target of the property access
has already been resolved. It's only necessary to resolve the property
itself, which can be done by calling
`PropertyElementResolver.resolvePropertyAccess` directly.

This change helps pave the way for a follow-up CL in which I'll be
adding assertions that fire if the `ResolverVisitor` tries to resolve
a given subexpression multiple times. (Doing so is dangerous because
it can interfere with the correctness of flow analysis.)

Change-Id: I41c0aaa75b4e7902d584ce36bd0a87874f0eaf8f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/400040
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>