[analyzer] avoid doing redundant work when rewriting a prefixed identifier.
When the `ResolverVisitor` rewrites an expression from a prefixed
identifier to a property access (which it only does for record types),
it's not necessary to re-resolve the target of the property access,
because it was already resolved before deciding to rewrite the
expression.

This change helps pave the way for a follow-up CL in which I'll be
adding assertions that fire if the `ResolverVisitor` tries to resolve
a given subexpression multiple times. (Doing so is dangerous because
it can interfere with the correctness of flow analysis.)

Change-Id: I429e0b744683e662772474967188bb1457168795
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/399886
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>